move network module to target
