+++ Modules/NetworkModule/Project.swift
@@ -1,0 +1,17 @@
+//
+//  LoginModule.swift
+//  BuddiesIOSManifests
+//
+//  Created by Anıl Taşkıran on 15.05.2025.
+//
+
+import Foundation
+import ProjectDescription
+import ProjectDescriptionHelpers
+
+let project = Project.module(
+    name: "Network",
+    dependencies: [
+        .buddiesNetwork
+    ]
+)
