+++ Project.swift
@@ -1,34 +1,9 @@
 import ProjectDescription
 import Foundation
-
-extension Target {
-    static func featureTarget(
-        name: String,
-        productName: String,
-        product: Product = .staticLibrary,
-        dependencies: [TargetDependency],
-        hasResources: Bool = false
-    ) -> Self {
-         target(
-            name: name,
-            destinations: .iOS,
-            product: .staticLibrary,
-            productName: productName,
-            bundleId: "com.swiftbuddies.\(productName.lowercased())",
-            deploymentTargets: .iOS("17.0"),
-            sources: ["SwiftBuddiesIOS/Targets/\(name)Module/Sources/**"],
-            resources: hasResources ? ["SwiftBuddiesIOS/Targets/\(name)Module/Resources/**"] : [],
-            dependencies: dependencies)
-    }
-}
+import ProjectDescriptionHelpers
 
 let project = Project(
     name: "Buddies",
-    packages: [
-        .remote(url: "https://github.com/google/GoogleSignIn-iOS.git", requirement: .exact("7.0.0")),
-        .remote(url: "https://github.com/apple/swift-argument-parser.git", requirement: .exact("1.3.0")),
-        .remote(url: "https://github.com/darkbringer1/BuddiesNetwork.git", requirement: .upToNextMajor(from: .init(0, 0, 1)))
-    ],
     targets: [
         .target(
             name: "SwiftBuddiesIOS",
@@ -61,193 +36,35 @@ let project = Project(
                 ]
             ),
             dependencies: [
-                .package(product: "GoogleSignIn", type: .runtime, condition: .none),
-                .package(product: "BuddiesNetwork", type: .runtime, condition: .none),
-                .target(Modules.design.target),
-                .target(Modules.auth.target),
-                .target(Modules.onboarding.target),
-                .target(Modules.login.target),
-                .target(Modules.feed.target),
-                .target(Modules.map.target),
-                .target(Modules.profile.target),
-                .target(Modules.contributors.target),
-                .target(Modules.network.target),
-                .target(Modules.localization.target),
-                .target(Modules.core.target)
+                .external(name: "GoogleSignIn"),
+                .external(name: "BuddiesNetwork"),
+                .target(Module.design.target),
+                .target(Module.auth.target),
+                .target(Module.login.target),
+                .target(Module.onboarding.target),
+                .target(Module.feed.target),
+                .target(Module.map.target),
+                .target(Module.profile.target),
+                .target(Module.contributors.target),
+                .target(Module.localization.target),
+                .target(Module.core.target),
+                Module.network.targetDependency,
             ]
         ),
-        Modules.design.target,
-        Modules.auth.target,
-        Modules.onboarding.target,
-        Modules.login.target,
-        Modules.feed.target,
-        Modules.map.target,
-        Modules.profile.target,
-        Modules.contributors.target,
-        Modules.network.target,
-        Modules.localization.target,
-        Modules.core.target,
-        Modules.localicationCodegen,
-        Modules.colorPaletteCodegen
+        Module.design.target,
+        Module.auth.target,
+        Module.onboarding.target,
+        Module.login.target,
+        Module.feed.target,
+        Module.map.target,
+        Module.profile.target,
+        Module.contributors.target,
+        Module.localization.target,
+        Module.core.target,
+        Module.localicationCodegen,
+        Module.colorPaletteCodegen
     ]
 )
-
-
-enum Modules: CaseIterable {
-    case core
-    case localization
-    case design
-    case network
-    case auth
-    case onboarding
-    case login
-    case feed
-    case map
-    case profile
-    case contributors
-    
-    
-    var target: Target {
-        switch self {
-        case .core:
-            Target.featureTarget(
-                name: "Core",
-                productName: "Core",
-                dependencies:
-                    [.target(Modules.auth.target), .target(Modules.network.target),
-                        .package(product: "GoogleSignIn", type: .runtime, condition: .none)]
-            )
-        case .localization:
-            Target.featureTarget(
-                name: "Localization",
-                productName: "Localization",
-                dependencies: [],
-                hasResources: true
-            )
-        case .design:
-            Target.featureTarget(
-                name: "Design",
-                productName: "Design",
-                dependencies: [.target(Modules.localization.target)],
-                hasResources: true
-            )
-        case .network:
-            Target.featureTarget(
-                name: "Network",
-                productName: "Network",
-                dependencies: [
-                    .package(product: "BuddiesNetwork", type: .runtime, condition: .none)
-                ]
-            )
-        case .auth:
-            Target.featureTarget(
-                name: "Auth",
-                productName: "Auth",
-                dependencies: [
-                    .target(Modules.network.target),
-                    .package(product: "GoogleSignIn", type: .runtime, condition: .none)
-                ]
-            )
-        case .onboarding:
-            Target.featureTarget(
-                name: "Onboarding",
-                productName: "Onboarding",
-                dependencies: [
-                    .target(Modules.design.target),
-                    .target(Modules.core.target),
-                ]
-            )
-        case .login:
-            Target.featureTarget(
-                name: "Login",
-                productName: "Login",
-                dependencies: [
-                    .target(Modules.design.target),
-                    .target(Modules.auth.target),
-                    .target(Modules.network.target),
-                    .target(Modules.core.target),
-                    .package(product: "GoogleSignIn", type: .runtime, condition: .none)
-                ]
-            )
-        case .feed:
-            Target.featureTarget(
-                name: "Feed",
-                productName: "Feed",
-                dependencies: [
-                    .target(Modules.core.target),
-                    .target(Modules.design.target),
-                    .package(product: "GoogleSignIn", type: .runtime, condition: .none)
-                ]
-            )
-        case .map:
-            Target.featureTarget(
-                name: "Map",
-                productName: "Map",
-                dependencies: [
-                    .target(Modules.core.target),
-                    .target(Modules.design.target)
-                ]
-            )
-        case .profile:
-            Target.featureTarget(
-                name: "Profile",
-                productName: "Profile",
-                dependencies: [
-                    .target(Modules.design.target),
-                    .target(Modules.auth.target),
-                    .target(Modules.network.target),
-                    .target(Modules.core.target),
-                    .package(product: "GoogleSignIn", type: .runtime, condition: .none)
-                ]
-            )
-        case .contributors:
-            Target.featureTarget(
-                name: "Contributors",
-                productName: "Contributors",
-                dependencies: [
-                    .target(Modules.core.target),
-                    .target(Modules.design.target)
-                ]
-            )
-        }
-    }
-    
-    static let localicationCodegen = Target.target(
-        name: "LocalizationCodegen",
-        destinations: .macOS,
-        product: .commandLineTool,
-        productName: "LocalizationCodegen",
-        bundleId: "com.swiftbuddies.localization",
-        sources: ["SwiftBuddiesIOS/Targets/ScriptsModule/LocalizationCodegen/**"],
-        scripts: [],
-        dependencies: [.package(product: "ArgumentParser", type: .runtime, condition: .none)],
-        coreDataModels: [],
-        environmentVariables: [:],
-        launchArguments: [],
-        additionalFiles: [],
-        buildRules: [],
-        mergedBinaryType: .automatic,
-        mergeable: false
-    )
-    
-    static let colorPaletteCodegen = Target.target(
-        name: "ColorPaletteCodegen",
-        destinations: .macOS,
-        product: .commandLineTool,
-        productName: "ColorCodege",
-        bundleId: "com.swiftbuddies.color",
-        sources: ["SwiftBuddiesIOS/Targets/ScriptsModule/ColorPaletteCodegen/**"],
-        scripts: [],
-        dependencies: [.package(product: "ArgumentParser", type: .runtime, condition: .none)],
-        coreDataModels: [],
-        environmentVariables: [:],
-        launchArguments: [],
-        additionalFiles: [],
-        buildRules: [],
-        mergedBinaryType: .automatic,
-        mergeable: false
-    )
-}
 
 //let scriptsModule = Target.target(
 //    name: "Scripts",
