+++ Tuist/Package.swift
@@ -15,8 +15,8 @@ import PackageDescription
 let package = Package(
     name: "SwiftBuddiesIOS",
     dependencies: [
-        // Add your own dependencies here:
-        // .package(url: "https://github.com/Alamofire/Alamofire", from: "5.0.0"),
-        // You can read more about dependencies here: https://docs.tuist.io/documentation/tuist/dependencies
+        .package(url: "https://github.com/darkbringer1/BuddiesNetwork.git", from: "0.0.1"),
+        .package(url: "https://github.com/google/GoogleSignIn-iOS.git", from: "7.0.0"),
+        .package(url: "https://github.com/apple/swift-argument-parser.git", from: "1.3.0")
     ]
 )
