+++ Workspace.swift
@@ -1,0 +1,15 @@
+//
+//  Workspace.swift
+//  BuddiesIOSManifests
+//
+//  Created by Anıl Taşkıran on 15.05.2025.
+//
+
+import ProjectDescription
+
+let workspace = Workspace(
+    name: "Buddies",
+    projects: [
+        "."
+    ]
+)
